refactor: extract axis classes from data.py into new axes.py module
- Create new src/pyhs3/axes.py module with all axis classes:
  - Base Axis class
  - UnbinnedAxis for unbinned data
  - RegularAxis and IrregularAxis for binned data
  - BinnedAxis discriminated union wrapper
  - BinnedAxes, UnbinnedAxes, and Axes collection classes
  - binned_axis_discriminator function

- Update src/pyhs3/data.py:
  - Remove axis class definitions
  - Import Axis, BinnedAxis, UnbinnedAxis from pyhs3.axes

- Update source file imports to use new module:
  - src/pyhs3/distributions/histfactory/__init__.py
  - src/pyhs3/distributions/histfactory/samples.py
  - src/pyhs3/distributions/histogram.py
  - src/pyhs3/functions/standard.py

This separation provides better module organization by isolating axis
specifications from data implementations.
